Move PDB and CIF parsing to their own classes in formats/
Register them for automatic file detection.
