+++ chemistry/formats/__init__.py
@@ -5,3 +5,4 @@ formats
 
 from chemistry.formats.registry import load_file
 from chemistry.formats import io
+from chemistry.formats.pdb import PDBFile, CIFFile
