+++ setup.py
@@ -9,7 +9,7 @@ packages = ['ParmedTools', 'ParmedTools.gui', 'ParmedTools.simulations']
 # Next the main chemistry packages
 packages += ['chemistry', 'chemistry.amber', 'chemistry.modeller',
              'chemistry.tinker', 'chemistry.unit', 'chemistry.amber.mdin',
-             'chemistry.charmm', 'chemistry.openmm', 'chemistry.pdbx',
+             'chemistry.charmm', 'chemistry.openmm', 'chemistry.formats.pdbx',
              'chemistry.formats', 'fortranformat']
 
 # Modules
